Fixed an assert that was being generated by mis-matched argument within a wxLogDebug() call at line 17758 in ExportUtilities.cpp's RebuildSourceText() function. The size_t type needed an (int) cast to agree with the %d specifier.

+++ source/ExportFunctions.cpp
@@ -17755,7 +17755,7 @@ int RebuildSourceText(wxString& source, SPList* pUseThisList)
 	source.Trim(); // trim the end
 	gpApp->GetDocument()->m_bCurrentlyFiltering = FALSE; // restore default, BEW 28Mar23
 #if defined(_DEBUG)
-	wxLogDebug(_T("Rebuild SRC: line %d, returning Length() = %d , source= [%s]"), __LINE__, source.Length(), source.c_str());
+	wxLogDebug(_T("Rebuild SRC: line %d, returning Length() = %d , source= [%s]"), __LINE__, (int)source.Length(), source.c_str());
 #endif
 
 	// update length
